use u64

+++ src/data/point.rs
@@ -1,11 +1,11 @@
 #[derive(Debug, Clone, Eq, PartialEq, Hash)]
 pub struct Point {
-    pub(crate) x: u128,
-    pub(crate) y: u128,
+    pub(crate) x: u64,
+    pub(crate) y: u64,
 }
 
 impl Point {
-    pub fn new(x: impl Into<u128>, y: impl Into<u128>) -> Self {
+    pub fn new(x: impl Into<u64>, y: impl Into<u64>) -> Self {
         Self {
             x: x.into(),
             y: y.into(),
@@ -15,8 +15,8 @@ impl Point {
 
 impl<K, V> From<(K, V)> for Point
 where
-    K: Into<u128>,
-    V: Into<u128>,
+    K: Into<u64>,
+    V: Into<u64>,
 {
     fn from((x, y): (K, V)) -> Self {
         Self::new(x, y)
@@ -25,8 +25,8 @@ where
 
 impl<K, V> From<&(K, V)> for Point
 where
-    K: Into<u128> + Copy,
-    V: Into<u128> + Copy,
+    K: Into<u64> + Copy,
+    V: Into<u64> + Copy,
 {
     fn from(pair: &(K, V)) -> Self {
         Self::new(pair.0, pair.1)
