+++ src/okvs/okvs.rs
@@ -1,5 +1,5 @@
 use crate::data::point::Point;
 
 pub trait OKVS {
-    fn decode(&self, key: impl Into<u128>) -> Point;
+    fn decode(&self, key: impl Into<u64>) -> Point;
 }
